+++ elephant/test/test_spike_train_synchrony.py
@@ -12,7 +12,7 @@ from quantities import Hz, ms, second
 import elephant.spike_train_generation as stgen
 from elephant.spike_train_synchrony import Synchrotool, spike_contrast, \
     _get_theta_and_n_per_bin, _binning_half_overlap
-from elephant.data_utils import download_elephant_data, unzip
+from elephant.datasets import download_datasets, unzip
 
 
 class TestSpikeContrast(unittest.TestCase):
@@ -166,8 +166,8 @@ class TestSpikeContrast(unittest.TestCase):
 
         izhikevich_gin = r"dataset-3/Data_Izhikevich_network.zip"
         checksum = "70e848500c1d9c6403b66de8c741d849"
-        filepath_zip = download_elephant_data(repo_path=izhikevich_gin,
-                                              checksum=checksum)
+        filepath_zip = download_datasets(repo_path=izhikevich_gin,
+                                         checksum=checksum)
         unzip(filepath_zip)
         filepath_json = filepath_zip.with_suffix(".json")
         with open(filepath_json) as read_file:
